+++ src/idunno.Bluesky/Video/BlueskyAgent.cs
@@ -15,49 +15,6 @@ public partial class BlueskyAgent
     private const string UploadBlobLxm = "com.atproto.repo.uploadBlob";
 
     /// <summary>
-    /// Gets the status details for the specified video processing job.
-    /// </summary>
-    /// <param name="jobId">The job id whose status should be queried.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentException">Thrown when <paramref name="jobId"/> is <see langword="null"/> or whitespace.</exception>
-    public async Task<AtProtoHttpResult<JobStatus>> GetVideoJobStatus(
-        string jobId,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentException.ThrowIfNullOrWhiteSpace(jobId);
-
-        using (_logger.BeginScope($"Getting jobStatus for {jobId}"))
-        {
-            AtProtoHttpResult<JobStatus> result = await BlueskyServer.GetVideoJobStatus(
-                jobId,
-                _videoServer,
-                HttpClient,
-                LoggerFactory,
-                cancellationToken: cancellationToken).ConfigureAwait(false);
-
-            if (result.Succeeded)
-            {
-                Logger.GetJobStatusSucceeded(_logger, jobId, result.Result.State, result.Result.Progress);
-            }
-            else
-            {
-                string? error = null;
-                string? message = null;
-                if (result.AtErrorDetail is not null)
-                {
-                    error = result.AtErrorDetail.Error;
-                    message = result.AtErrorDetail.Message;
-                }
-
-                Logger.GetJobStatusFailed(_logger, result.StatusCode, error, message);
-            }
-
-            return result;
-        }
-    }
-
-    /// <summary>
     /// Gets any video upload restrictions placed on the current user 
     /// </summary>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
@@ -75,7 +32,7 @@ public partial class BlueskyAgent
             AtProtoHttpResult<ServiceCredential> getServiceAuthResult = await GetServiceAuth(
                 service: Service,
                 audience: WellKnownDistributedIdentifiers.Video,
-                lxm: "app.bsky.video.getUploadLimits",
+                lxm: "com.atproto.repo.uploadBlob",
                 cancellationToken: cancellationToken).ConfigureAwait(false);
 
             if (!getServiceAuthResult.Succeeded)