+++ src/idunno.Bluesky/Video/Model/FinishUploadRequest.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+namespace idunno.Bluesky.Video.Model;
+
+internal record FinishUploadRequest(
+    string jobId,
+    JobStatusWireFormat jobStatus)
+{
+}