+++ src/idunno.Bluesky/Video/Model/GetUploadStatusResponse.cs
@@ -1,0 +1,19 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace idunno.Bluesky.Video.Model;
+
+internal record GetUploadStatusResponse(
+    [property: JsonRequired] string JobId,
+    [property: JsonRequired, JsonPropertyName("partSizeBytes")] long PartSize,
+    [property: JsonRequired] long PartCount,
+    [property: JsonRequired] IReadOnlyCollection<int> ReceivedParts,
+    [property: JsonRequired] DateTimeOffset ExpiresAt,
+    [property: JsonRequired] string State,
+    string? CompletedJobId,
+    string? JobStatus,
+    string? FailureReason)
+{
+}